base_station/core/requirements.txt
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
av==16.0.1
cffi==2.0.0
cryptography==46.0.3
dnspython==2.8.0
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
google-crc32c==1.7.1
grpcio==1.76.0
grpcio-tools==1.76.0
ifaddr==0.2.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
protobuf==6.33.1
psycopg==3.2.12
psycopg-binary==3.2.12
# psycopg2==2.9.11
pycparser==2.23
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
setuptools==80.9.0
typing_extensions==4.15.0
